// packages/lark/src/types.ts
export type ThreadItem = {
  message_id?: string | undefined;
  root_id?: string | undefined;
  parent_id?: string | undefined;
  thread_id?: string | undefined;
  msg_type?: string | undefined;
  create_time?: string | undefined;
  update_time?: string | undefined;
  deleted?: boolean | undefined;
  updated?: boolean | undefined;
  chat_id?: string | undefined;
  sender?:
    | {
        id: string;
        id_type: string;
        sender_type: string;
        tenant_key?: string | undefined;
      }
    | undefined;
  body?:
    | {
        content: string;
      }
    | undefined;
  mentions?:
    | {
        key: string;
        id: string;
        id_type: string;
        name: string;
        tenant_key?: string | undefined;
      }[]
    | undefined;
  upper_message_id?: string | undefined;
};

// blow code generated by gpt 😄
type Style = 'bold' | 'underline' | 'italic';

interface TextTag {
  tag: 'text';
  text: string;
  style: Style[];
}

interface ATag {
  tag: 'a';
  href: string;
  text: string;
  style: Style[];
}

interface AtTag {
  tag: 'at';
  user_id: string;
  user_name: string;
  style: Style[];
}

interface ImgTag {
  tag: 'img';
  image_key: string;
}

interface MediaTag {
  tag: 'media';
  file_key: string;
  image_key: string;
}

interface EmotionTag {
  tag: 'emotion';
  emoji_type: 'SMILE';
}

interface TextContent {
  text: string;
}

export type Content =
  | {
      title: string;
      content: (TextTag | ATag | AtTag | ImgTag | MediaTag | EmotionTag)[][];
    }
  | TextContent;